layout/generic/test/test_key_enter_open_second_summary.html

<!DOCTYPE html>
<html>
<head>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <script>
  SimpleTest.waitForExplicitFinish();
  function runTest() {
    // Dispatch 'return' key on second summary should not collapse details.
    const summary = document.getElementById("summary");
    summary.focus();
    synthesizeKey("KEY_Enter");

    const details = document.querySelector("details");
    ok(details.open, "Dispatch 'return' key on second summary should not collapse details.");

    SimpleTest.finish();
  }
  </script>
</head>
<body onload="runTest();">
  <details open>
    <summary>Summary</summary>
    <summary id="summary">Second Summary</summary>
    <p>This is the details.</p>
  </details>
</body>
</html>
